Updated styles, component names and layout
