--- packages/react-google-maps-api-gatsby-example/src/components/bootstrap-custom-forms.css
@@ -1,8 +1,0 @@
-.custom-control-label::before {
-    background-color: #fff;
-    border: #adb5bd solid 1px;
-}
-
-.custom-control-input:checked~.custom-control-label::before {
-    border-color: #007bff;
-}
