+++ packages/react-google-maps-api-gatsby-example/src/components/checkbox-bicycling.js
@@ -17,8 +17,6 @@ const CheckboxBicycling = ({ onChange, value }) => (
       value={value}
     />
 
-    &nbsp;
-
     <label
       className='custom-control-label'
       htmlFor={id}
