+++ packages/react-google-maps-api-gatsby-example/src/components/checkbox-ground.js
@@ -17,8 +17,6 @@ const CheckboxGround = ({ onChange, value }) => (
       value={value}
     />
 
-    &nbsp;
-
     <label
       className='custom-control-label'
       htmlFor={id}
