+++ packages/react-google-maps-api-gatsby-example/src/components/checkbox-shapes.js
@@ -17,8 +17,6 @@ const CheckboxShapes = ({ onChange, value }) => (
       value={value}
     />
 
-    &nbsp;
-
     <label
       className='custom-control-label'
       htmlFor={id}
