+++ packages/react-google-maps-api-gatsby-example/src/components/checkbox-traffic.js
@@ -17,8 +17,6 @@ const CheckboxTraffic = ({ onChange, value }) => (
       value={value}
     />
 
-    &nbsp;
-
     <label
       className='custom-control-label'
       htmlFor={id}
