+++ packages/react-google-maps-api-gatsby-example/src/components/header.js
@@ -2,17 +2,19 @@ import React from 'react'
 import { Link } from 'gatsby'
 import logo from '../img/logo.svg'
 
-import { logoStyle } from '../components/styles'
-
 const Header = () => (
   <div className='container'>
-    <div className='py-4 text-center'>
-      <Link to='/'>
-        <img src={logo} style={logoStyle} alt='Logo' />
-      </Link>
+    <div className='row'>
+      <div className='col-lg-10 offset-lg-1'>
+        <div className='header'>
+          <Link to='/'>
+            <img src={logo} className='logo' alt='Logo' />
+          </Link>
 
-      <h1 className='h2'>React Google Maps API</h1>
-      <p className='lead text-secondary'>Gatsby Example</p>
+          <h1 className='h3'>React Google Maps API</h1>
+          <p className='lead text-secondary'>Gatsby Example</p>
+        </div>
+      </div>
     </div>
   </div>
 )
