+++ packages/react-google-maps-api-gatsby-example/src/components/no-maps.js
@@ -1,0 +1,11 @@
+import React from 'react'
+
+const NoMaps = () => (
+  <div className='no-maps card shadow-sm'>
+    <div className='card-body'>
+      <p className='text-secondary my-0'>Please enter your Google API key to load map examples.</p>
+    </div>
+  </div>
+)
+
+export default NoMaps
