+++ packages/react-google-maps-api-gatsby-example/src/components/section-bicycling.js
@@ -3,21 +3,16 @@ import React from 'react'
 import PropTypes from 'prop-types'
 import { connect } from 'react-redux'
 
-import BicyclingExample from '../examples/bicycling-example'
+import ExampleBicycling from '../examples/example-bicycling'
 
-import {
-  mapStyle,
-  shapeExampleStyles
-} from '../components/styles'
+import { shapeExampleStyles } from '../components/styles'
 
 const SectionBicycling = ({ bicycling }) =>
   bicycling
     ? (
-      <div style={mapStyle}>
-        <BicyclingExample
-          styles={shapeExampleStyles}
-        />
-      </div>
+      <ExampleBicycling
+        styles={shapeExampleStyles}
+      />
     )
     : null
