+++ packages/react-google-maps-api-gatsby-example/src/components/section-language.js
@@ -5,8 +5,6 @@ import langEs from '../img/es.svg'
 import langRu from '../img/ru.svg'
 import { connect } from 'react-redux'
 
-import { flagStyle } from '../components/styles'
-
 import {
   changeLanguage
 } from '../actions/app'
@@ -37,65 +35,59 @@ class SectionLanguage extends Component {
   }
 
   render = () => (
-    <div className='row'>
-      <div className='col-sm-4 col-md-3 col-xl-2'>
-        <div className='custom-control custom-radio mb-1'>
-          <input
-            type='radio'
-            id='en'
-            className='custom-control-input'
-            checked={this.props.language === 'en'}
-            onChange={this.setEnLang}
-          />
-          {` `}
-          <label
-            className='custom-control-label'
-            htmlFor='en'
-          >
-            <img style={flagStyle} src={langEn} alt='EN' />
-            English
-          </label>
-        </div>
+    <div className='d-flex flex-wrap'>
+      <div className='form-group custom-control custom-radio mr-4'>
+        <input
+          type='radio'
+          id='en'
+          className='custom-control-input'
+          checked={this.props.language === 'en'}
+          onChange={this.setEnLang}
+        />
+        {` `}
+        <label
+          className='custom-control-label'
+          htmlFor='en'
+        >
+          <img src={langEn} className='flag' alt='EN' />
+          English
+        </label>
       </div>
 
-      <div className='col-sm-4 col-md-3 col-xl-2'>
-        <div className='custom-control custom-radio mb-1'>
-          <input
-            type='radio'
-            id='es'
-            className='custom-control-input'
-            checked={this.props.language === 'es'}
-            onChange={this.setEsLang}
-          />
-          {` `}
-          <label
-            className='custom-control-label'
-            htmlFor='es'
-          >
-            <img style={flagStyle} src={langEs} alt='ES' />
-            Spanish
-          </label>
-        </div>
+      <div className='form-group custom-control custom-radio mr-4'>
+        <input
+          type='radio'
+          id='es'
+          className='custom-control-input'
+          checked={this.props.language === 'es'}
+          onChange={this.setEsLang}
+        />
+        {` `}
+        <label
+          className='custom-control-label'
+          htmlFor='es'
+        >
+          <img src={langEs} className='flag' alt='ES' />
+          Spanish
+        </label>
       </div>
 
-      <div className='col-sm-4 col-md-3 col-xl-2'>
-        <div className='custom-control custom-radio mb-1'>
-          <input
-            type='radio'
-            id='ru'
-            className='custom-control-input'
-            checked={this.props.language === 'ru'}
-            onChange={this.setRuLang}
-          />
-          {` `}
-          <label
-            className='custom-control-label'
-            htmlFor='ru'
-          >
-            <img style={flagStyle} src={langRu} alt='RU' />
-            Russian
-          </label>
-        </div>
+      <div className='form-group custom-control custom-radio mr-4'>
+        <input
+          type='radio'
+          id='ru'
+          className='custom-control-input'
+          checked={this.props.language === 'ru'}
+          onChange={this.setRuLang}
+        />
+        {` `}
+        <label
+          className='custom-control-label'
+          htmlFor='ru'
+        >
+          <img src={langRu} className='flag' alt='RU' />
+          Russian
+        </label>
       </div>
     </div>
   )
