+++ packages/react-google-maps-api-gatsby-example/src/components/section-shapes.js
@@ -3,21 +3,16 @@ import React from 'react'
 import PropTypes from 'prop-types'
 import { connect } from 'react-redux'
 
-import ShapesExample from '../examples/shapes-example'
+import ExampleShapes from '../examples/example-shapes'
 
-import {
-  mapStyle,
-  shapeExampleStyles
-} from '../components/styles'
+import { shapeExampleStyles } from './styles'
 
 const SectionShapes = ({ shapes }) =>
   shapes
     ? (
-      <div style={mapStyle}>
-        <ShapesExample
-          styles={shapeExampleStyles}
-        />
-      </div>
+      <ExampleShapes
+        styles={shapeExampleStyles}
+      />
     )
     : null
