+++ packages/react-google-maps-api-gatsby-example/src/components/styles.js
@@ -1,47 +1,8 @@
 import indexStyles from '../pages/index.module.css'
 
-export const logoStyle = {
-  width: '96px',
-  height: '96px'
-}
-
-export const flagStyle = {
-  border: '1px solid #eee',
-  width: '22px',
-  height: '17px',
-  lineHeight: 0,
-  marginRight: '6px',
-  position: 'relative',
-  top: '-2px'
-}
-
-export const mapStyle = {
-  width: '100%',
-  height: 'auto',
-  maxWidth: '800px',
-  marginTop: '1rem'
-}
-
-export const btnStyle = {
-  padding: '.375rem 1.25rem'
-}
-
-export const footerStyle = {
-  backgroundColor: '#e9ecef',
-  borderTop: '1px solid #dee2e6',
-  marginTop: '3rem',
-  padding: '1rem 0',
-  textAlign: 'center'
-}
-
-export const loadingStyle = {
-  height: '100%',
-  backgroundColor: '#023456'
-}
-
 export const mapContainerStyle = {
   height: '400px',
-  width: '800px'
+  width: '100%'
 }
 
 export const shapeExampleStyles = {
