+++ packages/react-google-maps-api-gatsby-example/src/examples/example-heatmap.js
@@ -1,0 +1,39 @@
+import React from 'react'
+import PropTypes from 'prop-types'
+import {
+  GoogleMap,
+  HeatmapLayer
+} from '@react-google-maps/api'
+
+const ExampleHeatmapPropTypes = {
+  styles: PropTypes.shape({
+    container: PropTypes.object.isRequired
+  }).isRequired
+}
+
+const center = {
+  lat: 40.75,
+  lng: -74
+}
+
+const ExampleHeatmap = ({ styles }) => (
+  <div className='map'>
+    <div className='map-container'>
+      <GoogleMap
+        id='heatmap-example'
+        mapContainerStyle={styles.container}
+        zoom={10}
+        center={center}
+        onClick={(...args) => {
+          console.log('onClick args: ', args)
+        }}
+      >
+        <HeatmapLayer />
+      </GoogleMap>
+    </div>
+  </div>
+)
+
+ExampleHeatmap.propTypes = ExampleHeatmapPropTypes
+
+export default ExampleHeatmap
