+++ packages/react-google-maps-api-gatsby-example/src/pages/index.module.css
@@ -1,4 +1,1 @@
-.mapContainer {
-  border-radius: 0.5rem;
-  overflow: hidden;
-}
+
